Opal report metric values in opal-linear.csv. This permits plots of portfolio size vs. expected CE, which is useful for determining if over/underfunded.

+++ opal/src/com/gordoni/opal/VitalStats.java
@@ -428,7 +428,7 @@ public class VitalStats
                 }
         }
 
-        public double metric_divisor(MetricsEnum metric, int age)
+        public double metric_divisor(MetricsEnum metric, double age)
         {
                 if (metric != MetricsEnum.INHERIT)
                         return 1;
